extract ffi components out of index.js
commit-id:5fa573fd
